cleanup dead benchmark code
