+++ jailbreaks/benchmarks/__init__.py
@@ -1,4 +1,3 @@
 from .base import Benchmark
-from .refusal import RefusalBenchmark
 
-__all__ = ["Benchmark", "RefusalBenchmark"]
+__all__ = ["Benchmark"]
