+++ jailbreaks/benchmarks/base.py
@@ -1,11 +1,10 @@
-from jailbreaks.llm import LLM
+from typing import List
 
-class Benchmark:
-    def __init__(self, name=None):
-        self.name = name or self.__class__.__name__
+class Benchmark:    
+    def __init__(self, prompts: List[str] = None, max_new_tokens: int = 100, name: str = "refusal"):
+        self.name = name or self.__class__.__name__ 
+        self.prompts = prompts
+        self.max_new_tokens = max_new_tokens
     
-    def evaluate(self, model: LLM):
-        raise NotImplementedError
-    
-    def __str__(self):
-        return self.name
+    def get_prompts(self):
+        return self.prompts
